requirements.txt
pip>=24.2
langchain==0.0.200  # Update to a compatible version (check for the latest stable version)
langchain-community==0.0.3
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21
Pillow==9.0.1
streamlit>=1.33    # Update to at least 1.33
openai==0.27.0
altair==4.2.2
tiktoken==0.3.0
huggingface-hub==0.15.1
sentence-transformers==2.2.2
